+++ test/activitProcessApi.spec.js
@@ -15,13 +15,11 @@ describe('Activiti Process Api', function () {
         this.authResponseBpmMock.get200Response();
 
         this.alfrescoJsApi = new AlfrescoApi({
-            username: 'admin',
-            password: 'admin',
             host: this.hostActiviti,
             provider: 'BPM'
         });
 
-        this.alfrescoJsApi.login().then(() => {
+        this.alfrescoJsApi.login('admin', 'admin').then(() => {
             done();
         });
     });