+++ test/contet.spec.js
@@ -12,12 +12,10 @@ describe('Content', function () {
 
         this.authResponseMock.get201Response();
         this.alfrescoJsApi = new AlfrescoApi({
-            username: 'admin',
-            password: 'admin',
             host: this.host
         });
 
-        this.alfrescoJsApi.login().then(() => {
+        this.alfrescoJsApi.login('admin', 'admin').then(() => {
             done();
         });
     });