+++ test/webScriptApi.spec.js
@@ -18,12 +18,10 @@ describe('WebScript', function () {
 
         this.authResponseMock.get201Response();
         this.alfrescoJsApi = new AlfrescoApi({
-            username: 'admin',
-            password: 'admin',
             host: this.host
         });
 
-        this.alfrescoJsApi.login().then((data) => {
+        this.alfrescoJsApi.login('admin', 'admin').then(() => {
             done();
         });
     });